feat: 메인페이지를 로그인페이지로 변경 (#60)
