+++ src/app/components/SideBar.tsx
@@ -29,7 +29,7 @@ function SideBar() {
   const handleLogout = () => {
     showAlert("정말 로그아웃하시겠습니까?", () => {
       logout();
-      router.push("/login");
+      router.push("/");
     });
   };
